Refactor settings to no longer require Pydantic models.
Environmental variables are fetched using the os library.
