+++ usps/main.py
@@ -15,7 +15,7 @@ from aiohttp import ClientSession, ClientError
 from pydantic import BaseModel
 from google.oauth2 import service_account
 from google.auth.transport.requests import Request
-from usps.config import settings
+from usps import settings
 from usps import transaction_log
 
 app = FastAPI()
